requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
albumentations>=1.3.0
matplotlib>=3.7.0
pillow>=10.0.0
pandas>=2.0.0
pyyaml>=6.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6